#11 Store updated sessions
