+++ CardGameDurak.Network/SingleGameController.cs
@@ -3,10 +3,10 @@
 namespace CardGameDurak.Network;
 internal class SingleGameController : IGameController
 {
-    private readonly IGamesCoordinator<IAwaitPlayer> _gameCoordinator;
+    private readonly IGameCoordinator<IAwaitPlayer> _gameCoordinator;
     private readonly IBot _bot;
     public SingleGameController(
-        IGamesCoordinator<IAwaitPlayer> coordinator,
+        IGameCoordinator<IAwaitPlayer> coordinator,
         IBot bot)
     {
         _bot = bot ?? throw new ArgumentNullException(nameof(bot));
